integration/pyproject.toml
[project]
name = "jaxboost-integration-tests"
version = "0.1.0"
description = "Integration tests for JAXBoost"
requires-python = ">=3.10,<3.13"
dependencies = [
    "jaxboost>=0.5.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "scikit-learn>=1.3.0",
    "pytest>=7.4.0",
]

[tool.pytest.ini_options]
testpaths = ["."]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow",
    "xgboost: marks tests that require XGBoost",
    "lightgbm: marks tests that require LightGBM",
]
